(svn r10374) [0.5] -Backport from trunk (r10333, r10336, r10337, r10345, 10346, 10368): - Fix: Waypoints could be renamed when you are not the owner (r10368) - Fix: The 'old' pathfinders (OPF and NPF) for road vehicles could not find a path when in a tunnel [FS#290] (r10345, r10346) - Fix: Only add the autoreplace menu when autoreplace actually knows about the group [FS#880] (r10337) - Fix: Signal state sometimes not properly set when the signal "pathfinder" reached the end of a line [FS#910] (r10336) - Fix: News messages were shown over the endgame/highscore windows [FS#943] (r10333)

+++ waypoint.c
@@ -316,13 +316,15 @@ int32 CmdRenameWaypoint(TileIndex tile, uint32 flags, uint32 p1, uint32 p2)
 
 	if (!IsValidWaypointID(p1)) return CMD_ERROR;
 
+	wp = GetWaypoint(p1);
+	if (!CheckTileOwnership(wp->xy)) return CMD_ERROR;
+
 	if (_cmd_text[0] != '\0') {
 		StringID str = AllocateNameUnique(_cmd_text, 0);
 
 		if (str == 0) return CMD_ERROR;
 
 		if (flags & DC_EXEC) {
-			wp = GetWaypoint(p1);
 			if (wp->string != STR_NULL) DeleteName(wp->string);
 
 			wp->string = str;
@@ -335,7 +337,6 @@ int32 CmdRenameWaypoint(TileIndex tile, uint32 flags, uint32 p1, uint32 p2)
 		}
 	} else {
 		if (flags & DC_EXEC) {
-			wp = GetWaypoint(p1);
 			if (wp->string != STR_NULL) DeleteName(wp->string);
 
 			MakeDefaultWaypointName(wp);
